refactor: make components and helper functions
