+++ src/Components/Card.jsx
@@ -1,0 +1,5 @@
+function Card({ video }) {
+  return <h4>{video.items[0].snippet.title}</h4>;
+}
+
+export { Card };
