+++ src/Components/Header.jsx
@@ -1,0 +1,10 @@
+function Header() {
+  return (
+    <header>
+      <h2> Yt-Dēta </h2>
+      <p> Your goto app for extracting the data of Youtube Videos </p>
+    </header>
+  );
+}
+
+export { Header };
